created seperate login logout routes
