+++ src/lib/stores/pkceLoginStore.ts
@@ -1,0 +1,2 @@
+export const keycloak = 'http://auth.openfoodfacts.localhost:5600/realms/open-products-facts';
+export	const pkceClientId = 'OFF_EXP_DEV';
